requirements.txt
Flask==2.0.1
Werkzeug==2.0.3
flask-cors==4.0.0
numpy==1.23.5
tensorflow==2.13.1
Pillow==9.0.0
gunicorn==20.1.0